Main

Description

Documentation

move :: [String] -> Int -> Char -> [String]

A função move vai receber o mapa, o jogador e o comando e vai verificar se o comando é válido e se o jogador existe.

adicionaBomba :: Int -> Int -> (Int, Int) -> [String] -> [String]

Quando o jogador usa o comando B vai adicionar uma bomba ao estado

segundaCoordenadaBomb :: String -> Int -> String

Vai descobrir a segunda coordenada duma string de bomba, por exemplo, segundaCoordenadasBomb "+ 3 5" 0 - devolve 5

primeiraCoordenadaBomb :: String -> Int -> String

Vai descobrir a primeira coordenada duma string de bomba, por exemplo, primeiraCoordenadasBomb "+ 3 5" 0 - devolve 3

moveEsquerda :: [String] -> Int -> (Int, Int) -> [String]

Altera a posicao do jogador no estado do jogo quando move para esquerda

moveDireita :: [String] -> Int -> (Int, Int) -> [String]

Altera a posicao do jogador no estado do jogo quando move para direita

moveCima :: [String] -> Int -> (Int, Int) -> [String]

Altera a posicao do jogador no estado do jogo quando move para cima

moveBaixo :: [String] -> Int -> (Int, Int) -> [String]

Altera a posicao do jogador no estado do jogo quando move para baixo

converteString :: Int -> (Int, Int) -> String

Converte a posicao do jogador para string, por exemplo, converteString 0 (3,3) fica "0 3 3"

contaPowers :: String -> Int

Conta os powers que um jogador tem

desenhaPowers :: Int -> String -> String

Desenha os powers depois da posicao do jogador

verificaPower :: (Int, Int) -> [(Int, Int)] -> Bool

Vai verificar se o jogador tem algum power

adicionaPower :: [String] -> Int -> [String]

Vai aidicionar um power ao jogador, caso o jogador mover para uma posição e nessa posição tiver um power

verificaMapaEsquerda :: Int -> Int -> (Int, Int) -> [String] -> Bool

Verifica se pode mover para a esquerda

verificaMapaDireita :: Int -> Int -> (Int, Int) -> [String] -> Bool

Verifica se pode mover para a direita

verificaMapaCima :: Int -> Int -> (Int, Int) -> [String] -> Bool

Verifica se pode mover para cima

verificaMapaBaixo :: Int -> Int -> (Int, Int) -> [String] -> Bool

Verifica se pode mover para baixo

arranjaCoordenadasPowersUps :: [String] -> [String]

Percorre o estado de jogo todo e guarda a linha que começar por +

apagaMais :: [String] -> [String]

Apaga o +

apagaUmEspaco :: [String] -> Int -> [String]

Apaga um espaço

primeiraCoordenada :: String -> String

Devolve a 1ª coordenada da string, neste caso a coluna. (Ex: primeiraCoordenada "2 3" -> 2)

segundaCoordenada :: String -> Int -> String

Devolve a 2ª coordenada da string, neste caso a linha. (Ex: segundaCoordenada "2 3" -> 3)

listaPowerUps :: [String] -> [(Int, Int)]

Dá a lista de Power Ups sob forma de coordenada

listaPowerDireita :: [String] -> [(Int, Int)]

Dá a lista de Powers Ups no formato de coordenada sem o + e espaço.

apagaPower :: [String] -> String -> [String]

Apaga o power do estado do jogo

condicao :: String -> String -> Bool

Função auxiliar para o apagaPower

bombasColocadas :: [String] -> [String]

Percorre o estado do jogo e guarda a linha se começar por *

arranjaJogador :: [String] -> Int -> [Int]

Percorre a lista de bombas colocadas e guarda os jogadores que colocaram as respetivas bombas

verificaJogador :: Int -> [Int] -> Bool

Verifica se um determinado jogador já colocou bomba

converCoordsString :: (Int, Int) -> String

Converte as coordenadas para string (Ex: converCoordsString (2,3) -> "2 3"

arranjaCoordenadas :: Int -> [String] -> String

Percorre o estado de jogo e arranja a linha de coordenadas do respetivo jogador

coordenadas :: String -> Int -> Int -> [Int]

Percorre a linha do jogador e devolve a coluna e linha

coordY :: [Int] -> String

Converte a coordenada para o número original

coordX :: [Int] -> String

Converte a linha para o número original

verificaJogadorEst :: Int -> [String] -> Bool

Verifica se um jogador existe no estado de jogo

parCoordenadas :: String -> Int -> (Int, Int)

Converte a posição do jogador para coordenada

main :: IO ()

Função Main