Tarefa1

Description

Documentation

mapa :: Int -> Int -> [String]

’mapa’: Função Original de todas as funções, é a aquela que vai devolver o mapa mais os estados.

desenhaTodos :: Int -> Int -> String

Desenha todos os '#' numa linha, por exemplo, desenhaTodos 0 9 vai desenhar "#########"

desenhaInicFim :: Int -> Int -> String

Desenha os '#' na primeira/ultima posição, por exemplo, desenhaInicFim 0 9 vai desenhar ""

desenhaInicFim1 :: Int -> Int -> String

Desenha o '#' na primeiraultima posição, mas nesta função, quando é na posição 12(d-2)(d-3) desenha um '_' , assim sabe-se que nessas posições é onde o jogador começa. Por exemplo, desenhaInicFim1 0 9 vai desenhar ""

desenhaPosPares :: Int -> Int -> String

Desenha os '#' nas posições pares, por exemplo, desenhaPosPares 0 9 " #"

desenhaPosPares1 :: Int -> Int -> String

Desenha os '#' nas posições pares, por exemplo, desenhaPosPares1 0 9 " _#". Só que também vai adicionar um "_" de forma a prevenir que o jogador possa mover para cima/baixo (Dependendo do jogador), pois se houver um '#' o jogador não se pode mover

mapaOriginal :: Int -> Int -> [String]

Desenhar o mapa só com '#' e '_' , por exemplo, mapaOriginal 0 9 vai devolver ["######## _ ",""," _ ","","#########"] utilizando as funções auxiliares definidas.

desenhaPontos :: [Int] -> String -> String

Desenhar os "?" numa linha

desenhaPontosMapa :: Int -> Int -> [Int] -> [String] -> [String]

Vai desenhar todos os ? que existem no mapa.

coordsBombFlares :: Int -> Int -> [Int]

0-1 - Power Ups Bombs escondidos atrás de um tijolo; 2-3 - Power Ups Flames escondidos atrás de um tijolo; 4-39 - Tijolo 40-99 - Vazio

contaEspacos :: String -> Int

É preciso contar espaços livres Numa linha para depois apagar da lista de números

contaEspacosMapa :: [String] -> Int

Contar todos os espaços livres no mapa

apaga :: String -> String

Apagar os traços numa linha apenas

apagaTracosMapa :: [String] -> [String]

Apaga todos os traços que existirem no mapa

desenhaBombMap :: Int -> Int -> Int -> [Int] -> [String] -> [String]

A função vai percorrer o mapa e a lista de números ao mesmo tempo, quando achar 0/1 na lista e for um ? no mapa, então vai desenhar uma bomba no fim do mapa.

desenhaFlameMap :: Int -> Int -> Int -> [Int] -> [String] -> [String]

A função vai percorrer o mapa e a lista de números ao mesmo tempo, quando achar 2/3 na lista e for um ? no mapa, então vai desenhar um flame no fim do mapa.

contaLinha :: String -> Int

Conta o nº de espaços livres ou ? existente uma linha.