Main

Description

Documentation

encode :: [String] -> String

Codifica o estado do jogo para uma string

dimensao :: [String] -> String

Determina a dimensão do mapa

contaLinhas :: [String] -> Int

Conta o número de linhas que começam por '#' de forma a saber a dimensão

dimensaoString :: Int -> String

Converte a dimensão de inteiro para string

posicaoBombsFlames :: Int -> Int -> Int -> [String] -> String

Percorre o mapa e à medida que encontra ? guarda a respetiva posição

posicaoJogador :: [String] -> String

Guarda a posição de cada jogador numa string

bombsFlames :: [String] -> String

Guarda as informações das bombs/flames numa string

detetaBombasColocadas :: [String] -> String

Deteta as bombas colocadas pelos jogadores e guarda numa string

decode :: String -> [String]

Descodifica a string para o estado de jogo

parCoordenadas :: String -> (Int, Int)

Converte uma string com par de coordenadas, por exemplo "2 3", para um par de inteiros, neste caso (2,3)

primeiraCoordenada :: String -> String

Devolve a 1ª coordenada da string, neste caso a coluna. (Ex: primeiraCoordenada "2 3" -> 2)

segundaCoordenada :: String -> Int -> String

Devolve a 2ª coordenada da string, neste caso a linha. (Ex: segundaCoordenada "2 3" -> 3)

listaPosicoes :: [String] -> [(Int, Int)]

Converte as coordenadas da string para uma lista de coordenadas. (Ex: listaPosicoes ["2 3","15 3"] -> [(2,3),(15,3)]

getPosicoes :: String -> Int -> String

Dá a posição de cada jogador numa string

convertePosicoesString :: String -> String

Função auxiliar da função converteLista

converteLista :: String -> [String]

Converte as posições para uma lista de strings com posições

arranjarPosicoes :: String -> String

Arranja as posições de cada bomb/flame

apaga :: String -> String

Tira os traços no processo de conversão de posições para lista

getBombasColocadas :: String -> Int -> String

Obtém informação das posições dos jogadores numa determinada linha

getBombsFlames :: String -> Int -> String

Obtém informação das posições dos bomb/flames numa determinada linha

desenhaPontos :: [(Int, Int)] -> Int -> Int -> String -> String

Desenhar os "?" numa linha

apagaPontos :: [(Int, Int)] -> Int -> Int -> String -> [(Int, Int)]

Remove da lista os pontos,já colocados no mapa

desenhaPontosMapa :: [(Int, Int)] -> Int -> Int -> [String] -> [String]

Desenha os ? no mapa todo

getDimensao :: String -> String

Dá a dimensão do mapa

converteDimensao :: String -> Int

Converte a dimensão para inteiro

desenhaTodos :: Int -> Int -> String

desenhaInicFim :: Int -> Int -> String

desenhaPosPares :: Int -> Int -> String

mapaOriginal :: Int -> Int -> [String]

main :: IO ()

Função main